Add 5-model sweep config and auto-dtype handling
- ModelSpec gains dtype and no_think fields; each model declares its preferred dtype
- load_model falls back to spec.dtype when caller omits dtype argument
- apply_chat_template passes enable_thinking=False for Qwen3 (no_think=True models)
- phi3 updated to TL-native loader; new entries: qwen3 (1.7B), olmo3-7b, gemma3-1b, mistral7b
- run_experiment.py drops hardcoded float32 so spec.dtype takes effect automatically
- FINDINGS.md next-steps section expanded with full model sweep table

+++ kaggle/run_experiment.py
@@ -72,7 +72,7 @@ OUT      = pathlib.Path("/kaggle/working")
 
 spec   = MODELS[MODEL]
 device = "cuda" if torch.cuda.is_available() else "cpu"
-loaded = load_model(spec, device=device, dtype="float32")
+loaded = load_model(spec, device=device)  # dtype comes from ModelSpec.dtype
 print(f"Loaded {spec.key}: {loaded.n_layers}L × {loaded.n_heads}H  on {device}")
 
 harm  = load_advbench(limit=N_PAIRS * 4)
